+++ Frontend/lib/pages/profile-view.dart
@@ -4,278 +4,211 @@ import 'package:sprintf/sprintf.dart';
 import '../custom-widgets/custom-drawer.dart';
 import '../custom-widgets/custom-app-bar.dart';
 import '../util/appcolor.dart';
+import '../router.dart' as router;
+import '../custom-widgets/tab-list-view.dart';
 
 class ProfileView extends StatefulWidget {
-  ProfileView({Key key}) : super(key: key);
+  final bool isNoRequest;
+
+  ProfileView({Key key, this.isNoRequest}) : super(key: key);
 
   _ProfileViewState createState() => new _ProfileViewState();
 }
 
 class _ProfileViewState extends State<ProfileView> {
-  void navDrawBar() {}
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
         appBar: CustomAppBar(),
         endDrawer: CustomDrawer(),
         backgroundColor: AppColor.primaryColor,
-        body: Stack(
-          children: <Widget>[
-            SingleChildScrollView(
-              child: Stack(children: <Widget>[profileScreen()]),
-            )
-          ],
+        body: Container(
+          child: Column(mainAxisSize: MainAxisSize.min, children: <Widget>[
+            profileUserDetailsSection(),
+            Padding(padding: EdgeInsets.only(top: 40, bottom: 5)),
+            profileButtonSection(),
+            Padding(padding: EdgeInsets.only(top: 40, bottom: 5)),
+            myRequestsLabelSection(),
+            Padding(padding: EdgeInsets.only(top: 20, bottom: 5)),
+            Expanded(
+              child: myRequestsSection(),
+            ),
+          ]),
         ));
   }
 
-  Padding createText(text, topPadding) {
-    return Padding(
-      padding: EdgeInsets.fromLTRB(10, topPadding, 20, 5),
-      child: Container(
-        alignment: Alignment.topLeft,
-        child: Text(text,
-            style: TextStyle(
-                fontSize: 16,
-                color: AppColor.primaryDarkColor,
-                fontFamily: 'Montserrat',
-                fontWeight: FontWeight.w400)),
-      ),
+  Widget profileUserDetailsSection() {
+    return Column(
+      children: [
+        Padding(
+          padding: EdgeInsets.fromLTRB(30, 30, 20, 5),
+          child: Container(
+            alignment: Alignment.topLeft,
+            child: Text("Pavan",
+                style: TextStyle(
+                    fontSize: 20,
+                    color: AppColor.primaryDarkColor,
+                    fontFamily: 'Montserrat',
+                    fontWeight: FontWeight.bold)),
+          ),
+        ),
+        Padding(
+          padding: EdgeInsets.fromLTRB(30, 5, 20, 0),
+          child: Container(
+            alignment: Alignment.topLeft,
+            child: Text("#83, Church Street, Bengaluru",
+                style: TextStyle(
+                    fontSize: 13,
+                    color: AppColor.primaryDarkColor,
+                    fontFamily: 'Montserrat',
+                    fontWeight: FontWeight.w300)),
+          ),
+        )
+      ],
     );
   }
 
-  Widget createInputField(obscureText, keyboardType) {
-    return Container(
-      child: TextField(
-        keyboardType: keyboardType,
-        obscureText: obscureText,
-        decoration: InputDecoration(
-          filled: true,
-          fillColor: Colors.white,
-          contentPadding: const EdgeInsets.fromLTRB(20, 0, 20, 0),
-          border: OutlineInputBorder(
-              borderRadius: BorderRadius.all(Radius.circular(12))),
-        ),
-      ),
-    );
-  }
-
-  Widget profileScreen() {
-    return Center(
-        child: new Container(
-            padding: const EdgeInsets.only(right: 30.0, left: 30.0),
-            child: new Container(
-              child:
-                  new Center(child: new Column(children: getProfileSection())),
-            )));
-  }
-
-  List<Widget> getProfileSection() {
-    List<Widget> profileSectionList = List<Widget>();
-    profileSectionList.add(new Padding(padding: EdgeInsets.only(top: 0.0)));
-    //profileSectionList.addAll(profilePicSection());
-    profileSectionList.addAll(personalInfoSection());
-    profileSectionList.addAll(changePasswordSection());
-    profileSectionList.add(savePasswordBtn());
-    profileSectionList.add(new Padding(padding: EdgeInsets.only(bottom: 10.0)));
-    return profileSectionList;
-  }
-
-  /*List<Widget> profilePicSection() {
-    List<Widget> profilePicList = List<Widget>();
-    profilePicList.add(new Row(
-      children: <Widget>[
-        new Padding(padding: EdgeInsets.only(left: 10.0)),
-        Container(
-          child: Image.asset("assets/dpp.jpg"),
-          height: 160,
-          width: 120,
-        ),
-        new Padding(padding: EdgeInsets.only(right: 50.0)),
-        new InkWell(
-          onTap: () => print('edit avtar'),
-          child: new Container(
-            width: 150.0,
-            height: 60.0,
-            decoration: new BoxDecoration(
-              color: Color(0xFF232C4D),
-              border: new Border.all(color: Color(0xFF232C4D), width: 2.0),
-              borderRadius: new BorderRadius.circular(10.0),
+  Widget profileButtonSection() {
+    return Row(
+      children: [
+        Padding(
+          padding: EdgeInsets.fromLTRB(30, 0, 20, 5),
+          child: Container(
+            //alignment: Alignment.topLeft,
+            child: new InkWell(
+              onTap: () => editProfile(),
+              child: new Container(
+                width: 150.0,
+                height: 50.0,
+                decoration: new BoxDecoration(
+                  color: AppColor.primaryDarkColor,
+                  border: new Border.all(
+                      color: AppColor.primaryDarkColor, width: 2.0),
+                  borderRadius: new BorderRadius.circular(10.0),
+                ),
+                child: new Center(
+                  child: Row(
+                    children: [
+                      new Padding(padding: EdgeInsets.only(right: 10)),
+                      new IconTheme(
+                        data: new IconThemeData(color: AppColor.fontColor),
+                        child: new Icon(Icons.supervised_user_circle),
+                      ),
+                      new Padding(padding: EdgeInsets.only(right: 10)),
+                      new Text('Edit Profile',
+                          style: TextStyle(
+                              fontSize: 16,
+                              color: AppColor.fontColor,
+                              fontFamily: 'Montserrat',
+                              fontWeight: FontWeight.bold))
+                    ],
+                  ),
+                ),
+              ),
             ),
-            child: Row(
-              // Replace with a Row for horizontal icon + text
-              children: <Widget>[
-                new Padding(padding: EdgeInsets.only(left:20, )),
-                Icon(IconData(58313, fontFamily: 'MaterialIcons'),color: Color(0xFFF5FBFF),),
-                new Padding(padding: EdgeInsets.only(left:10)),
-                new Center(
-                  child: new Text('Edit Avatar',
-                      style: new TextStyle(
-                          fontSize: 18.0, color: Color(0xFFF5FBFF))),
-                )
-              ],
+          ),
+        ),
+        Padding(
+          padding: EdgeInsets.fromLTRB(20, 0, 20, 5),
+          child: Container(
+            //alignment: Alignment.topLeft,
+            child: new InkWell(
+              onTap: () => addItem(),
+              child: new Container(
+                width: 150.0,
+                height: 50.0,
+                decoration: new BoxDecoration(
+                  color: AppColor.primaryDarkColor,
+                  border: new Border.all(
+                      color: AppColor.primaryDarkColor, width: 2.0),
+                  borderRadius: new BorderRadius.circular(10.0),
+                ),
+                child: new Center(
+                  child: Row(
+                    children: [
+                      new Padding(padding: EdgeInsets.only(right: 10)),
+                      new IconTheme(
+                        data: new IconThemeData(color: AppColor.fontColor),
+                        child: new Icon(Icons.add_circle),
+                      ),
+                      new Padding(padding: EdgeInsets.only(right: 10)),
+                      new Text('Add Item',
+                          style: TextStyle(
+                              fontSize: 16,
+                              color: AppColor.fontColor,
+                              fontFamily: 'Montserrat',
+                              fontWeight: FontWeight.bold))
+                    ],
+                  ),
+                ),
+              ),
             ),
           ),
         )
       ],
-    ));
-    return profilePicList;
-  }*/
-
-  List<Widget> personalInfoSection() {
-    List<Widget> personalInfoSectionList = List<Widget>();
-    /*personalInfoSectionList
-        .add(new Padding(padding: EdgeInsets.only(top: 30.0)));
-    personalInfoSectionList.add(getField("Name"));
-    personalInfoSectionList
-        .add(new Padding(padding: EdgeInsets.only(top: 30.0)));
-    personalInfoSectionList.add(getField("Email"));
-    personalInfoSectionList
-        .add(new Padding(padding: EdgeInsets.only(top: 30.0)));
-    personalInfoSectionList.add(getField("Address"));
-    personalInfoSectionList
-        .add(new Padding(padding: EdgeInsets.only(bottom: 40.0)));*/
-    personalInfoSectionList.add(createText("Name", 30.0));
-    personalInfoSectionList.add(createInputField(false, TextInputType.text));
-    personalInfoSectionList.add(createText("Email", 30.0));
-    personalInfoSectionList.add(createInputField(false, TextInputType.emailAddress));
-    personalInfoSectionList.add(createText("Address", 30.0));
-    personalInfoSectionList.add(createInputField(false, TextInputType.multiline));
-    personalInfoSectionList
-        .add(new Padding(padding: EdgeInsets.only(bottom: 40.0)));
-    return personalInfoSectionList;
-  }
-
-  List<Widget> changePasswordSection() {
-    List<Widget> changePasswordSectionList = List<Widget>();
-    changePasswordSectionList.add(changePasswordLabel());
-    changePasswordSectionList
-        .add(new Padding(padding: EdgeInsets.only(bottom: 5.0)));
-    changePasswordSectionList.add(Container(
-        height: 5,
-        decoration: new BoxDecoration(
-          color: AppColor.primaryDarkColor,
-          border: new Border.all(color: AppColor.primaryDarkColor, width: 0.0),
-          borderRadius: new BorderRadius.circular(25.0),
-        ),
-        child: Divider(thickness: 1, color: AppColor.primaryDarkColor)));
-    /*changePasswordSectionList
-        .add(new Padding(padding: EdgeInsets.only(bottom: 50.0)));
-    changePasswordSectionList.add(getField("Current Password"));
-    changePasswordSectionList
-        .add(new Padding(padding: EdgeInsets.only(top: 30.0)));
-    changePasswordSectionList.add(getField("New Password"));
-    changePasswordSectionList
-        .add(new Padding(padding: EdgeInsets.only(top: 30.0)));
-    changePasswordSectionList.add(getField("Confirm New Password"));*/
-    changePasswordSectionList.add(createText("Current Password", 30.0));
-    changePasswordSectionList.add(createInputField(true, null));
-    changePasswordSectionList.add(createText("New Password", 30.0));
-    changePasswordSectionList.add(createInputField(true, null));
-    changePasswordSectionList.add(createText("Confirm New Password", 30.0));
-    changePasswordSectionList.add(createInputField(true, null));
-    changePasswordSectionList
-        .add(new Padding(padding: EdgeInsets.only(bottom: 40.0)));
-    return changePasswordSectionList;
-  }
-
-  Widget changePasswordLabel() {
-    return new Container(
-        child: new Text(
-      'Change Password',
-      style: new TextStyle(
-          fontWeight: FontWeight.normal,
-          fontSize: 22.0,
-          color: AppColor.primaryDarkColor),
-    ));
-  }
-
-  Widget savePasswordBtn() {
-    return new InkWell(
-      onTap: () => processProfileForm(),
-      child: new Container(
-        width: 250.0,
-        height: 70.0,
-        decoration: new BoxDecoration(
-          color: AppColor.primaryDarkColor,
-          border: new Border.all(color: AppColor.primaryDarkColor, width: 2.0),
-          borderRadius: new BorderRadius.circular(10.0),
-        ),
-        child: new Center(
-          child: new Text('Save Changes',
-              style: new TextStyle(fontSize: 18.0, color: AppColor.fontColor)),
-        ),
-      ),
     );
   }
 
-  Widget getField(fieldName) {
-    TextInputType fieldTextInputType = null;
-    bool obscureText = false;
-    switch (fieldName) {
-      case "Email":
-        fieldTextInputType = TextInputType.emailAddress;
-        break;
-      case "Address":
-        fieldTextInputType = TextInputType.multiline;
-        break;
-      case "Current Password":
-        fieldTextInputType = TextInputType.visiblePassword;
-        obscureText = true;
-        break;
-      case "Confirm New Password":
-        fieldTextInputType = TextInputType.visiblePassword;
-        obscureText = true;
-        break;
-      case "New Password":
-        fieldTextInputType = TextInputType.visiblePassword;
-        obscureText = true;
-        break;
-      default:
-        fieldTextInputType = TextInputType.text;
-        break;
-    }
-
-    Widget textField = TextFormField(
-      obscureText: obscureText,
-      decoration: new InputDecoration(
-        focusColor: AppColor.fontColor,
-        enabledBorder: OutlineInputBorder(
-          borderSide: BorderSide(color: AppColor.primaryDarkColor, width: 1.0),
-          borderRadius: BorderRadius.circular(25.0),
+  Widget myRequestsLabelSection() {
+    return Column(
+      children: [
+        Container(
+          child: new Text('My Requests',
+              style: TextStyle(
+                  fontSize: 16,
+                  color: AppColor.primaryDarkColor,
+                  fontFamily: 'Montserrat',
+                  fontWeight: FontWeight.bold)),
         ),
-        focusedBorder: OutlineInputBorder(
-          borderSide: BorderSide(color: AppColor.primaryDarkColor, width: 1.0),
-          borderRadius: BorderRadius.circular(25.0),
-        ),
-        labelStyle:
-            new TextStyle(color: AppColor.primaryDarkColor, fontSize: 16.0),
-        labelText: sprintf("%s", [fieldName]),
-        fillColor: AppColor.fontColor,
-        filled: true,
-        border: new OutlineInputBorder(
-          borderSide: const BorderSide(
-              color: Colors.white, width: 2.0, style: BorderStyle.solid),
-          borderRadius: BorderRadius.circular(25.0),
-        ),
-        //fillColor: Colors.green
-      ),
-      validator: (val) {
-        if (val.length == 0) {
-          return sprintf("%s cannot be empty", [fieldName]);
-        } else {
-          return null;
-        }
-      },
-      keyboardType: fieldTextInputType,
-      style: new TextStyle(
-        fontFamily: "Poppins",
-      ),
+        new Padding(padding: EdgeInsets.only(bottom: 10.0)),
+        Container(
+          height: 4,
+          width: 350,
+          decoration: new BoxDecoration(
+            color: AppColor.primaryDarkColor,
+            border:
+                new Border.all(color: AppColor.primaryDarkColor, width: 0.0),
+            borderRadius: new BorderRadius.circular(25.0),
+          ),
+        )
+      ],
     );
-    return textField;
   }
 
-  void processProfileForm() {
+  Widget myRequestsSection() {
+    return Container(
+        width: MediaQuery.of(context).size.width - 20,
+        child: Container(
+            width: 350,
+            decoration: new BoxDecoration(
+                color: const Color(0xFFF5FBFF),
+                borderRadius: new BorderRadius.only(
+                    topLeft: const Radius.circular(10.0),
+                    topRight: const Radius.circular(10.0))),
+            height: MediaQuery.of(context).size.height,
+            child: widget.isNoRequest != null
+                ? getNoRequestSection()
+                : TabListView()));
+  }
+
+  Widget getNoRequestSection() {
+    return Container(
+        alignment: Alignment.topCenter,
+        padding: EdgeInsets.only(top: 50.0),
+        child: Text("You haven't requested for\n any commodity yet!",
+            style: TextStyle(
+                letterSpacing: 1,
+                fontSize: 18,
+                color: AppColor.primaryDarkColor,
+                fontFamily: 'Montserrat',
+                fontWeight: FontWeight.w500)));
+  }
+
+  void addItem() {
     print("hello!");
+  }
+
+  void editProfile() {
+    Navigator.pushReplacementNamed(this.context, router.PROFILE_EDIT);
   }
 }
