+++ Frontend/lib/util/constants.dart
@@ -8,5 +8,6 @@ const INITIAL_ROUTE = '/';
 const REGISTRATION = 'REGISTRATION';
 const THANKYOU = 'THANKYOU'; // Thank you for signing up Page
 const EMAIL_CONFIRMATION = 'EMAIL_CONFIRMATION';
-const PROFILE = 'PROFILE';
+const PROFILE_EDIT = 'PROFILE_EDIT';
+const PROFILE_VIEW = 'PROFILE_VIEW';
 const ABOUT_US = 'ABOUT US';
