+++ Frontend/pubspec.yaml
@@ -64,14 +64,16 @@ flutter:
         - asset: fonts/Pacifico-Regular.ttf
     - family: Montserrat
       fonts:
-        - asset: fonts/Montserrat-Black.ttf
+        - asset: fonts/Montserrat/Montserrat-Black.ttf
           weight: 900
         - asset: fonts/Montserrat-Regular.ttf
           weight: 400
         - asset: fonts/Montserrat-Bold.ttf
           weight: 600
-        - asset: fonts/Montserrat-Medium.ttf
+        - asset: fonts/Montserrat/Montserrat-Medium.ttf
           weight: 500
+        - asset: fonts/Montserrat/Montserrat-Italic.ttf
+          weight: 300
 
     # - family: Trajan Pro
     #   fonts:
